SystemUI: make long pressing recent switch to last app
With this patch, long pressing the recents key on the navigation bar
will attempt to switch to your last application in the activity stack
before the active one.

The user's current home launcher package is ignored as well as SystemUI (so that
RecentsActivity does not influence the behavior).

Change-Id: I38771e0fce16b55bb5186af47115b4e812cb60b0
Signed-off-by: Roman Birg <roman@cyngn.com>
Signed-off-by: Adnan Begovic <adnan@cyngn.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Long pressing recent switch to last app (Remove Custom animation)

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	packages/SystemUI/res/anim/last_app_in.xml
	packages/SystemUI/res/anim/last_app_out.xml

Change-Id: I0eea3b1bf4bbcb6feebd3756a308db05d3fcf4df
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>